Added start of grid rework #1
